make gae-sessions thread-safe
Previously, a global variable was used to store the session associated with the
current request.  This was sufficient in the past (and to date) but sometime in
the future app engine is going to support threading.  The minor change
introduced by this commit ensures that gae-sessions will continue to work
properly even when being used by multiple threads.